backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic-settings
python-multipart
bcrypt
